requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
opencv-python==4.10.0.84
scikit-learn==1.5.1
tensorflow==2.16.1
keras==3.4.1
pillow 
streamlit

